Fix system manager unit tests
