+++ .github/workflows/unit-tests-system-manager.yml
@@ -5,13 +5,13 @@ on:
     branch:
       - 'main'
   pull_request:
-    paths:
-      - 'nuvlaedge/broker/'
-      - 'nuvlaedge/common/'
-      - 'nuvlaedge/models/'
-      - 'nuvlaedge/peripherals/'
-      - 'nuvlaedge/system_manager/'
-      - '.github/workflows/unit-tests-system-manager.yml'
+#    paths:
+#      - 'nuvlaedge/broker/'
+#      - 'nuvlaedge/common/'
+#      - 'nuvlaedge/models/'
+#      - 'nuvlaedge/peripherals/'
+#      - 'nuvlaedge/system_manager/'
+#      - '.github/workflows/unit-tests-system-manager.yml'
   workflow_dispatch:
